+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 
 - 支持 Flow, 类型检查
 - Breadcrumb 新增支持 separator , 自定义分隔符
+- Modal autoResizeHeight 属性重命名为 overflow
+- Button shape 属性 重命名为 appearance
 - 新增组件
   - Table
+  - Drawer
+  - Progress